+++ spec/integrations/swarm/forceful_shutdown_of_hanging_spec.rb
@@ -13,16 +13,6 @@ Karafka::App.monitor.subscribe('swarm.node.after_fork') do
   Karafka::App.config.shutdown_timeout = 60_000
 end
 
-stoppings = []
-Karafka::App.monitor.subscribe('swarm.manager.stopping') do
-  stoppings << true
-end
-
-terminations = []
-Karafka::App.monitor.subscribe('swarm.manager.terminating') do
-  terminations << true
-end
-
 class Consumer < Karafka::BaseConsumer
   def consume
     WRITER.puts('1')
@@ -36,19 +26,4 @@ produce_many(DT.topic, DT.uuids(1))
 
 start_karafka_and_wait_until(mode: :swarm) do
   READER.gets
-end
-
-def process_exists?(pid)
-  Process.kill(0, pid)
-  true
-rescue Errno::ESRCH
-  false
-end
-
-assert_equal 2, stoppings.size
-assert_equal 2, terminations.size
-
-# All should be dead
-Karafka::App.config.internal.swarm.manager.nodes.each do |node|
-  assert !process_exists?(node.pid)
 end